Move `mangle_directory_names()` -> `datalad_dataverse.utils`
